hooks fixes

+++ src/modules/lessons/hooks/useLessonsCall.ts
@@ -2,45 +2,55 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { addLessonsCall, deleteLessonsCall, updateLessonsCall } from '../api/api';
 import { LessonPlanRequest } from '../types/lessons';
 import { useNavigate } from 'react-router-dom';
+import { useRefreshTokenCall } from '../../auth/useRefreshTokenCall';
+import { toast } from 'react-toastify';
+import { Messages } from '../../../shared/constants/messages';
 
 export const useLessonsCall = () => {
+  const { ensureValidAccessToken } = useRefreshTokenCall();
   const queryClient = useQueryClient();
   const navigate = useNavigate();
 
   const { mutateAsync: addLessonsMutation } = useMutation({
     mutationKey: ['lessonsSpecification'],
-    mutationFn: async (lessons: LessonPlanRequest) => await addLessonsCall(lessons),
-    // onError: () => Toaster here,
-    // onSuccess: data => Toaster here,
+    mutationFn: async (lessons: LessonPlanRequest) => {
+      await ensureValidAccessToken();
+      return await addLessonsCall(lessons);
+    },
+    onError: () => toast(Messages.ERROR),
+    onSuccess: () => {
+      navigate('/lessons');
+      queryClient.invalidateQueries({ queryKey: ['lessonsList'] });
+      toast(Messages.ADDED);
+    },
   });
 
   const { mutateAsync: deleteLessonsMutation } = useMutation({
     mutationKey: ['lessonsSpecification'],
     mutationFn: async (id: string) => {
-      // await ensureValidAccessToken();
+      await ensureValidAccessToken();
       return await deleteLessonsCall(id);
     },
-    // onError: () => Toaster here,
+    onError: () => toast(Messages.ERROR),
     onSuccess: () => {
       navigate('/lessons');
+      toast(Messages.DELETED);
       return queryClient.invalidateQueries({ queryKey: ['lessonsList'] });
-
-      // Toaster here
     },
   });
 
   const { mutateAsync: updateLessonsMutation } = useMutation({
     mutationKey: ['vehicleSpecification'],
     mutationFn: async ({ id, lessons }: { id: string; lessons: LessonPlanRequest }) => {
-      // await ensureValidAccessToken();
+      await ensureValidAccessToken();
       return await updateLessonsCall(id, lessons);
     },
-    // onError: () => Toaster here,
+    onError: () => toast(Messages.ERROR),
     onSuccess: data => {
       navigate('/lessons/' + data._id);
-      return queryClient.invalidateQueries({ queryKey: ['lessonsDetails'] });
+      toast(Messages.UPDATED);
 
-      // Toaster here
+      return queryClient.invalidateQueries({ queryKey: ['lessonsDetails'] });
     },
   });
