+++ src/modules/news/hooks/useNewsCall.ts
@@ -1,18 +1,21 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { NewsRequest } from '../types/news';
 import { addNewsCall, deleteNewsCall, updateNewsCall } from '../api/api';
-// import { useRefreshTokenCall } from '../../auth/hooks/useRefreshTokenCall';
+import { useRefreshTokenCall } from '../../auth/useRefreshTokenCall';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { Messages } from '../../../shared/constants/messages';
 
 export const useNewsCall = () => {
-  // const { ensureValidAccessToken } = useRefreshTokenCall();
+  const { ensureValidAccessToken } = useRefreshTokenCall();
   const queryClient = useQueryClient();
   const navigate = useNavigate();
   const { mutateAsync: addNewsMutation } = useMutation({
-    mutationKey: ['vehicleSpecification'],
-    mutationFn: async (news: NewsRequest) => await addNewsCall(news),
+    mutationKey: ['newsSpecification'],
+    mutationFn: async (news: NewsRequest) => {
+      await ensureValidAccessToken();
+      return await addNewsCall(news);
+    },
     onError: () => toast(Messages.ERROR),
     onSuccess: () => {
       navigate('/news');
@@ -22,31 +25,30 @@ export const useNewsCall = () => {
   });
 
   const { mutateAsync: deleteNewsMutation } = useMutation({
-    mutationKey: ['vehicleSpecification'],
+    mutationKey: ['newsSpecification'],
     mutationFn: async (id: string) => {
-      // await ensureValidAccessToken();
+      await ensureValidAccessToken();
       return await deleteNewsCall(id);
     },
-    // onError: () => Toaster here,
+    onError: () => toast(Messages.ERROR),
     onSuccess: () => {
       navigate('/news');
+      toast(Messages.DELETED);
       return queryClient.invalidateQueries({ queryKey: ['newsList'] });
-
-      // Toaster here
     },
   });
 
   const { mutateAsync: updateNewsMutation } = useMutation({
-    mutationKey: ['vehicleSpecification'],
+    mutationKey: ['newsSpecification'],
     mutationFn: async ({ id, news }: { id: string; news: NewsRequest }) => {
-      // await ensureValidAccessToken();
+      await ensureValidAccessToken();
       return await updateNewsCall(id, news);
     },
-    // onError: () => Toaster here,
+    onError: () => toast(Messages.ERROR),
     onSuccess: data => {
       navigate('/news/' + data._id);
+      toast(Messages.UPDATED);
       return queryClient.invalidateQueries({ queryKey: ['newsDetails'] });
-      // Toaster here
     },
   });
